AutoGen/requirements.txt
# Python 3.10+ required (autogen-ext needs 3.10+)
autogen-agentchat>=0.2.36
autogen-ext[openai]>=0.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
